requirements.txt
scikit_image==0.17.2
python_bidi==0.4.2
six==1.15.0
matplotlib==2.2.3
scipy==1.1.0
torch==1.2.0
numba==0.39.0
Shapely==1.7.1
numpy==1.18.5
Flask==1.0.2
tensorflow_gpu==1.15.0
opencv_python==4.4.0.44
pathlib2==2.3.2
torchvision==0.4.0
easyocr==1.2.2
Pillow==8.1.0
plumbum==1.6.9
skimage==0.0
tensorflow==1.15.0
